dashboard/backend/requirements.txt
flask>=3.1,<4.0
flask-cors>=5.0,<6.0
ruamel.yaml>=0.18
